Feature/add srp principle (#24)
* fix error in filter method


* Update form action in filter


* Add a new mapping in the Books controller


* add index page


* Add functionality for editing and saving books

In 'FormEditBook' servlet, retrieve book details for editing and prepare data for the 'FormEditBook.jsp' page.

In 'SaveBook' servlet, handle book saving functionality. Extract parameters, create a new Book instance, and save it. Redirect to the 'ViewBook' servlet to display the updated book list.


* Mejoras en redirecciones


* Add styling to the book form


* update


---------

+++ Aplication01/src/main/java/com/arquitecturajava/ControlerBook.java
@@ -19,7 +19,17 @@ public class ControlerBook extends HttpServlet {
 		System.out.println(request.getServletPath());
 
 		if (request.getServletPath().equals("/ControlerBook/ViewBook")) {
-			List<Book> listaDeLibros = Book.searchAll();
+			List<Book> listaDeLibros = null;
+			String selectdISBN = request.getParameter("selectedCategory");
+			System.out.println("existe selectedCtegory : ");
+			System.out.println(request.getParameter("selectedCategory"));
+			if (request.getParameter("selectedCategory") == null)
+				selectdISBN = "seleccionar";
+			if (selectdISBN.equals("seleccionar")) {
+				listaDeLibros = Book.searchAll();
+			} else {
+				listaDeLibros = Book.searchForCategory(selectdISBN);
+			}
 			System.out.println(listaDeLibros.isEmpty());
 			List<String> listaDeCategorias = Book.searchAllCategory();
 			System.out.println(listaDeCategorias.isEmpty());
@@ -38,11 +48,26 @@ public class ControlerBook extends HttpServlet {
 			var libro = new Book(isbn, titulo, categoria);
 			libro.insert();
 			despachador = request.getRequestDispatcher("/ControlerBook/ViewBook");
-		} else {
+		} else if (request.getServletPath().equals("/ControlerBook/FormEditBook")) {
+			String isbn = request.getParameter("isbn");
+			List<String> listaDeCategorias = Book.searchAllCategory();
+			Book book = Book.findByIsbn(isbn);
+			request.setAttribute("libro", book);
+			request.setAttribute("listaDeCategorias", listaDeCategorias);
+			despachador = request.getRequestDispatcher("/FormEditBook.jsp");
+		} else if(request.getServletPath().equals("/ControlerBook/DeleteBook")) {
 			String isbn = request.getParameter("isbn");
 			var book = new Book(isbn);
 			book.delete();
 			despachador = request.getRequestDispatcher("/ControlerBook/ViewBook");
+		}else if(request.getServletPath().equals("/ControlerBook/SaveBook")) {
+			String isbn = request.getParameter("isbn");
+			String titulo = request.getParameter("titulo");
+			String categoria = request.getParameter("categoria");
+			var libro = new Book(isbn, titulo, categoria);
+			libro.save();
+			despachador = request.getRequestDispatcher("/ControlerBook/ViewBook");
+			
 		}
 		despachador.forward(request, response);
 	}
